\name{make.a.Matrix}
\alias{make.a.Matrix}
\title{Make a sparse matrix}
\usage{
  make.a.Matrix(data, nrow, ncol)
}
\arguments{
  \item{data}{the data to form into a matrix}

  \item{nrow}{the number of rows in the matrix}

  \item{ncol}{the number of columns in the matrix}
}
\value{
  the requested matrix
}
\description{
  Didactic function to make a sparse matrix
}
\details{
  I have no details for you.
}

