\name{transpose.a.Matrix}
\alias{transpose.a.Matrix}
\title{Take the transpose of a matrix you create}
\usage{
  transpose.a.Matrix(data, nrow, ncol)
}
\arguments{
  \item{data}{the data to form into a matrix}

  \item{nrow}{the number of rows in the ORIGINAL matrix}

  \item{ncol}{the number of columns in the ORIGINAL matrix}
}
\value{
  the transpose requested matrix
}
\description{
  Didactic function to make a sparse matrix and take its
  transpose
}
\details{
  I have no details for you.
}

